Add thing data, conversion, and meter 2D array to readingUtils and use these imported values in test C16
